+++ HvZ/Model/Domain/KillDomain.cs
@@ -11,8 +11,8 @@ namespace HvZ.Model.Domain
         public DateTime TimeOfDeath { get; set; }
         [MaxLength(50)]
         public string? Story { get; set; }
-        public float? Lat { get; set; }
-        public float? Lng { get; set; }
+        public double? Lat { get; set; }
+        public double? Lng { get; set; }
         [Required]
         public int GameId { get; set; }
         [Required]
